requirements.txt
streamlit==1.41.0
google-genai==1.14.0
python-dotenv==1.0.1
requests==2.32.3
